python/interview/pyproject.toml
[tool.poetry]
name = "oes.interview"
version = "0.2.0"
description = "OES Interview server"
authors = ["Wes <wes@anthronewengland.com>"]
license = "LGPL-3.0-or-later"
readme = "README.md"
packages = [{ include = "oes" }]

[tool.poetry.dependencies]
python = "^3.10"

attrs = "^23.1"
cattrs = "^23.1"
orjson = "^3.8.8"
typing-extensions = "^4.7.1"
pynacl = "^1.5.0"
jinja2 = "^3.1.2"
pyparsing = "^3.0.9"
importlib-metadata =  "^6.0.0"
loguru = "^0.7.0"
ruamel-yaml = "^0.18.5"
publicsuffixlist =  "^0.10.0.20230617"
email-validator =  "^2.0.0.post2"
httpx = { version = "^0.23.3" }
typed-settings = "^23.0.0"
blacksheep = "^1.2.18,!=1.2.19,!=1.2.20"
uvicorn = {extras = ["standard"], version = "^0.23.2"}

oes-util = { git = "https://github.com/Open-Event-Systems/registration.git", rev = "fb8f27f42774aa57c63d622de8cd9f4328adc2c4", subdirectory = "python/util" }
oes-template = { git = "https://github.com/Open-Event-Systems/registration.git", rev= "df218d2a0362b3c42f13b5ce5701e58a2f052f1c", subdirectory = "python/template" }

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
pytest-cov = "^4.0.0"
myst-parser = "^0.18.1"
sphinx = "^5.3.0"
sphinx-rtd-theme = "^1.1.1"
pytest-asyncio = "^0.21.0"

[tool.pytest.ini_options]
addopts = "--cov=oes.interview"
testpaths = [
    "tests",
]

[tool.pyright]
venvPath = "."
venv = ".venv"

# Console script entry point
[tool.poetry.plugins."console_scripts"]
oes-interview = "oes.interview.server.run:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
